requirements.txt
flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.2.1
requests==2.32.5
gunicorn==21.2.0
